+++ examples/resources/youtrack_project_custom_field/resource.tf
@@ -6,10 +6,11 @@ resource "youtrack_project" "parent" {
 
 # Attach the global "Priority" custom field to the project, using a project-specific bundle
 resource "youtrack_project_custom_field" "priority" {
-  project_id   = youtrack_project.parent.id
-  field_name   = "Priority"
-  field_type   = "EnumProjectCustomField"
-  bundle_name  = "My Project Priorities"
-  can_be_empty = true
-  is_public    = true
+  project_id          = youtrack_project.parent.id
+  field_name          = "Priority"
+  field_type          = "EnumProjectCustomField"
+  bundle_name         = "My Project Priorities"
+  default_value_names = ["Major"]
+  can_be_empty        = true
+  is_public           = true
 }